+++ src/Fable.AST/Plugins.fs
@@ -1,6 +1,5 @@
 namespace Fable
 
-open System
 open Fable.AST
 open Fable.AST.Fable
 
@@ -17,8 +16,4 @@ type PluginAttribute() =
 type MemberDeclarationPluginAttribute() =
     inherit PluginAttribute()
     abstract Transform: PluginHelper * MemberDecl -> MemberDecl
-
-[<AbstractClass>]
-type CallPluginAttribute() =
-    inherit PluginAttribute()
-    abstract Transform: PluginHelper * expr: Expr -> Expr
+    abstract TransformCall: PluginHelper * member_: MemberFunctionOrValue * expr: Expr -> Expr