+++ src/Fable.AST/RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 3.0.0-nagareyama-alpha-002
+
+* Merge Member/Call plugins
+
 ### 3.0.0-nagareyama-alpha-001
 
 * First alpha release